Add RESTful API and user authentication layer
This API is following the MVC pattern. Every Models, Routes and Controllers are
clearly linked with the same filename.
Data is persisted with MongoDB and Mongoose ODM.
Every user inputs are validated and sanitized before being stored in the
database.
Actions performing changes to the resources (POST, PUT, DELETE) are subjected
to the user being registered and logged-in.
Authentication is managed with JSON Web Token generated server-side and stored
on the user's computer.